// src/health.enum.ts
export enum Health {
  HTTP_PING = "HTTP_PING",
  HTTP_RES = "HTTP_RES",
  MONGOOSE = "MONGOOSE",
  MICROSERVICE = "MICROSERVICE",
  GRPC = "GRPC",
  MEMORY_RSS = "MEMORY_RSS",
  MEMORY_HEAP = "MEMORY_HEAP",
  DISK = "DISK",
  MIKRO_ORM = "MIKRO_ORM",
  SEQUELIZE = "SEQUELIZE",
  TYPE_ORM = "TYPE_ORM",
}